Update: Modularity of unit tests for helper functions
